+++ demo/djf_surveys
@@ -1,1 +1,1 @@
-/mnt/data/dev/django-form-surveys/djf_surveys
+/Users/apple/Documents/work_magang/django-form-surveys/djf_surveys/
