+++ djf_surveys/admins/views.py
@@ -18,14 +18,15 @@ from djf_surveys.app_settings import SURVEYS_ADMIN_BASE_PATH
 from djf_surveys.models import Survey, Question, UserAnswer
 from djf_surveys.mixin import ContextTitleMixin
 from djf_surveys.views import SurveyListView
-from djf_surveys.forms import BaseSurveyForm, CreateSurveyForm, SurveyWithChoicesForm
+from djf_surveys.forms import BaseSurveyForm
 from djf_surveys.summary import SummaryResponse
+from djf_surveys.admins.v2.forms import SurveyForm
 
 
 @method_decorator(staff_member_required, name='dispatch')
 class AdminCrateSurveyView(ContextTitleMixin, CreateView):
     template_name = 'djf_surveys/admins/form.html'
-    form_class = SurveyWithChoicesForm
+    form_class = SurveyForm
     title_page = _("Add New Survey")
 
     def post(self, request, *args, **kwargs):
@@ -42,7 +43,7 @@ class AdminCrateSurveyView(ContextTitleMixin, CreateView):
 @method_decorator(staff_member_required, name='dispatch')
 class AdminEditSurveyView(ContextTitleMixin, UpdateView):
     model = Survey
-    form_class = SurveyWithChoicesForm
+    form_class = SurveyForm
     template_name = 'djf_surveys/admins/form.html'
     title_page = _("Edit Survey")
